Moved generate_tp_model_with_activation_mp method for tests to generate_test_tp_model.py
