+++ tests/keras_tests/feature_networks_tests/feature_networks/mixed_precision_bops_test.py
@@ -17,7 +17,7 @@ from model_compression_toolkit import KPI, MixedPrecisionQuantizationConfigV2
 from keras.layers import Conv2D, Conv2DTranspose, DepthwiseConv2D, Dense, BatchNormalization, ReLU, Input, Add
 
 from model_compression_toolkit.core.common.mixed_precision.kpi_tools.kpi_methods import weights_size_kpi
-from tests.common_tests.helpers.activation_mp_tp_model import generate_tp_model_with_activation_mp
+from tests.common_tests.helpers.generate_test_tp_model import generate_tp_model_with_activation_mp
 from tests.keras_tests.feature_networks_tests.base_keras_feature_test import BaseKerasFeatureNetworkTest
 from model_compression_toolkit.core.tpc_models.default_tpc.latest import get_op_quantization_configs
 import tensorflow as tf
