+++ tests/pytorch_tests/model_tests/feature_models/mixed_precision_activation_test.py
@@ -19,7 +19,7 @@ from model_compression_toolkit import MixedPrecisionQuantizationConfig, KPI, Cor
     MixedPrecisionQuantizationConfigV2
 from model_compression_toolkit.core.common.user_info import UserInformation
 from model_compression_toolkit.core.tpc_models.default_tpc.latest import get_op_quantization_configs
-from tests.common_tests.helpers.activation_mp_tp_model import generate_tp_model_with_activation_mp
+from tests.common_tests.helpers.generate_test_tp_model import generate_tp_model_with_activation_mp
 from tests.pytorch_tests.model_tests.base_pytorch_test import BasePytorchTest
 import model_compression_toolkit as mct
 from tests.pytorch_tests.tpc_pytorch import get_mp_activation_pytorch_tpc_dict
