+++ tests/pytorch_tests/model_tests/feature_models/mixed_precision_bops_test.py
@@ -17,7 +17,7 @@ from model_compression_toolkit import MixedPrecisionQuantizationConfigV2, KPI
 from tests.pytorch_tests.model_tests.base_pytorch_test import BasePytorchTest
 
 from model_compression_toolkit.core.tpc_models.default_tpc.latest import get_op_quantization_configs
-from tests.common_tests.helpers.activation_mp_tp_model import generate_tp_model_with_activation_mp
+from tests.common_tests.helpers.generate_test_tp_model import generate_tp_model_with_activation_mp
 from tests.pytorch_tests.tpc_pytorch import generate_activation_mp_tpc_pytorch
